requirements.txt
flask==2.3.3
requests==2.31.0
transformers==4.45.2
torch==2.2.1
fpdf==1.7.2
Pillow==10.0.0 